로고 이미지, 앱 네임 변경

+++ jetpack_sample_source/step_tutorial/app/src/main/java/com/zahi/steptestapp/MainActivity.kt
@@ -37,7 +37,7 @@ class MainActivity : AppCompatActivity(), SensorEventListener {
 
     override fun onSensorChanged(event: SensorEvent) {
         if (event.sensor.type == Sensor.TYPE_STEP_COUNTER) {
-            binding.stepCount.text = event.values[0].toString()
+            binding.stepCount.text = event.values[0].toInt().toString()
         }
     }
